refactor: models 내 남은 파일 src 아래로 이동
