+++ src/apis/index.ts
@@ -1,0 +1,5 @@
+import * as GET from './methods/get';
+
+export const apiService = {
+  ...GET,
+};
