+++ src/config/index.ts
@@ -1,4 +1,11 @@
 import 'dotenv/config';
+import type { DataSourceOptions } from 'typeorm';
+import { ROOT_PATH } from '@/constants';
+import { join } from 'path';
+import { User } from '@/modules/user/entities';
+import { SubmitCode } from '@/modules/submit/entities';
+import { Problem } from '@/modules/problem/entities';
+import { TestCase } from '@/modules/problem/entities';
 
 class ConfigService {
   constructor(private env: { [key: string]: string | undefined }) {}
@@ -52,7 +59,7 @@ const requiredEnvironments = [
   'JWT_SECRET',
 ] as const;
 
-const configService = new ConfigService(process.env)
+export const configService = new ConfigService(process.env)
   .ensureValues(Object.values(requiredEnvironments), true)
   .ensureValues(
     Object.values(nonRequiredEnvironments),
@@ -61,4 +68,20 @@ const configService = new ConfigService(process.env)
   TupleToObject<typeof requiredEnvironments> &
   TupleToObject<typeof nonRequiredEnvironments>;
 
-export { configService };
+export const dataSourceOptions: DataSourceOptions = {
+  type: 'postgres',
+  host: configService.POSTGRE_HOST,
+  port: parseInt(configService.POSTGRE_PORT),
+  username: configService.POSTGRE_USERNAME,
+  password: configService.POSTGRE_PASSWORD,
+  database: configService.POSTGRE_DATABASE,
+  synchronize: !configService.isProduction(),
+  logging: !configService.isProduction(),
+  entities: [User, SubmitCode, Problem, TestCase],
+  migrations: [
+    configService.isProduction()
+      ? join(ROOT_PATH, 'src', 'models', 'migrations', '*.ts')
+      : '',
+  ],
+  subscribers: [],
+};
