+++ src/modules/auth/auth.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { OAuthApp } from 'octokit';
-import { configService } from '@/config/config.service';
+import { configService } from '@/config';
 import { BadRequestException } from '@nestjs/common';
 import { tokenService } from '@/utils';
 import { User } from '@/modules/user/entities';
